+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Cleanup westmere and haswell object files (#24) Thanks @fobser
+- Empty base16 and base64 in CDS and CDNSKEY can be represented with a '0'.
+  As specified in Section 4 of RFC 8078.
 
 ## [0.2.0] - 2024-12-12